mpu_statements_service/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.8
django-environ==0.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jws==0.1.3
jwt==1.0.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
